FleetManagement_CEO_Chatbot/requirements.txt
streamlit==1.34.0
pymongo==4.7.0
python-dotenv==1.0.1
google-generativeai==0.6.0  # For Gemini API
pandas==2.2.2
